# perf(core-database-postgres): add index on transactions.type