--- a/packages/core-database-postgres/src/migrations/20191008000000-add-type-index-to-transactions-table.sql
+++ b/packages/core-database-postgres/src/migrations/20191008000000-add-type-index-to-transactions-table.sql
@@ -1,0 +1,1 @@
+CREATE INDEX IF NOT EXISTS "transactions_type" ON ${schema~}.transactions ("type");